project directory restructure
