--- app/tests/aapl_tests.py
@@ -1,5 +1,0 @@
-import unittest
-
-
-
-
