+++ web_client/templates/layout.html
